+++ front/src/components/Modal.vue
@@ -5,19 +5,11 @@
         <div class="modal-container">
 
           <div class="modal-content">
-            <slot />
+            <slot name="content" />
           </div>
 
           <footer class="modal-footer">
-            <slot name="footer">
-              <button
-                class="modal-default-button float-right"
-                color="amber accent-4"
-                @click="$emit('close')"
-              >
-                디폴트: 등록
-              </button>
-            </slot>
+            <slot name="footer" />
           </footer>
         </div>
       </div>
@@ -52,9 +44,9 @@ export default {
 .modal-container {
   width: 80%;
   margin: 0px auto;
-  padding: 20px 30px;
+  padding: 10px 10px;
   background-color: #fff;
-  border-radius: 2px;
+  border-radius: 10px;
   box-shadow: 0 2px 8px rgba(0, 0, 0, 0.33);
   transition: all 0.3s ease;
   font-family: Helvetica, Arial, sans-serif;
