+++ front/src/utils/constants.js
@@ -1,5 +1,6 @@
 const ERROR_MESSAGE = {
   UNIDENTIFIABLE_LOCATION: "위치 정보를 확인할 수 없습니다.",
+  POST_CONTENT_NOT_FOUND: "내용을 입력해주세요.",
 };
 
 const API_BASE_URL = {
